+++ src/HoYo/Internal/Types.hs
@@ -7,6 +7,10 @@ Maintainer  : freddyjepringle@gmail.com
 Types used by all the main HoYo.* modules.
 -}
 
+{-# LANGUAGE DataKinds       #-}
+{-# LANGUAGE GADTs           #-}
+{-# LANGUAGE KindSignatures  #-}
+{-# LANGUAGE RankNTypes      #-}
 {-# LANGUAGE TemplateHaskell #-}
 {-# OPTIONS -Wno-missing-signatures #-}
 {-# OPTIONS_HADDOCK prune #-}
@@ -20,6 +24,7 @@ import Control.Monad.Trans.Reader (ReaderT)
 
 import Lens.Micro.TH
 
+import Data.List (intercalate)
 import qualified Data.Text as T
 import Data.Time
 
@@ -68,15 +73,50 @@ instance Show BookmarkSearchTerm where
   show (SearchIndex idx) = '#' : show idx
   show (SearchName name) = T.unpack name
 
+-- | The types of config values allowed in the HoYo config.
+data ConfigValueType =
+    TBool
+  | TDefaultBookmark
+  | TCommand
+  | TList ConfigValueType
+  | TMaybe ConfigValueType
+
+-- | Values in the HoYo config. Using a GADT parameterised by 'ConfigValueType'
+-- gives us stricter type safety.
+data ConfigValue (t :: ConfigValueType) where
+  BoolV             :: Bool -> ConfigValue 'TBool
+  DefaultBookmarkV  :: DefaultBookmark -> ConfigValue 'TDefaultBookmark
+  CommandV          :: T.Text -> ConfigValue 'TCommand
+
+  ListOfV           :: forall (a :: ConfigValueType) . [ConfigValue a] -> ConfigValue ('TList a)
+  MaybeV            :: forall (a :: ConfigValueType) . Maybe (ConfigValue a) -> ConfigValue ('TMaybe a)
+
+instance Show (ConfigValue (t :: ConfigValueType)) where
+  show (BoolV bool) = show bool
+  show (DefaultBookmarkV bm) = show bm
+  show (CommandV t) = show t
+  show (ListOfV xs) = "[" <> intercalate ", " (map show xs) <> "]"
+  show (MaybeV xs) = show xs
+
+instance Eq (ConfigValue (t :: ConfigValueType)) where
+  BoolV b1            == BoolV b2             = b1 == b2
+  DefaultBookmarkV b1 == DefaultBookmarkV b2  = b1 == b2
+  CommandV b1         == CommandV b2          = b1 == b2
+  ListOfV b1          == ListOfV b2           = b1 == b2
+  MaybeV b1           == MaybeV b2            = b1 == b2
+
+-- | Existential wrapper around 'ConfigValue'.
+data AnyConfigValue = forall (t :: ConfigValueType) . AnyConfigValue (ConfigValue t)
+
 -- | A representation of hoyo settings.
 data Config = Config {
-  _failOnError            :: !Bool
-  , _displayCreationTime  :: !Bool
-  , _enableClearing       :: !Bool
-  , _enableReset          :: !Bool
-  , _backupBeforeClear    :: !Bool
-  , _defaultBookmarks     :: ![DefaultBookmark]
-  , _defaultCommand       :: !(Maybe T.Text)
+  _failOnError            :: !(ConfigValue 'TBool)
+  , _displayCreationTime  :: !(ConfigValue 'TBool)
+  , _enableClearing       :: !(ConfigValue 'TBool)
+  , _enableReset          :: !(ConfigValue 'TBool)
+  , _backupBeforeClear    :: !(ConfigValue 'TBool)
+  , _defaultBookmarks     :: !(ConfigValue ('TList 'TDefaultBookmark))
+  , _defaultCommand       :: !(ConfigValue ('TMaybe 'TCommand))
   } deriving (Show, Eq)
 
 -- | 'HoYoMonad' is the main monad stack for the hoyo program. It's essentially a wrapper
@@ -100,5 +140,4 @@ data ExecResult =
 
 makeLenses ''Bookmark
 makeLenses ''DefaultBookmark
-makeLenses ''Config
 makeLenses ''Env