+++ src/HoYo/Internal/Utils.hs
@@ -7,27 +7,22 @@ Maintainer  : freddyjepringle@gmail.com
 Utility functions used by all the main HoYo.* modules.
 -}
 
-{-# LANGUAGE GADTs           #-}
-{-# LANGUAGE RankNTypes      #-}
-{-# LANGUAGE RecordWildCards #-}
+{-# LANGUAGE DataKinds  #-}
+{-# LANGUAGE GADTs      #-}
+{-# LANGUAGE RankNTypes #-}
+{-# OPTIONS_HADDOCK prune #-}
 
 module HoYo.Internal.Utils where
 
 import HoYo.Internal.Types
 
 import Data.Bifunctor (bimap, first)
-import Data.Char (isAscii, ord)
-import qualified Data.HashMap.Strict as HashMap
-import Data.List
-import qualified Data.List.NonEmpty as NE
-import Data.Semigroup (stimes)
 import qualified Data.Text as T
 import qualified Data.Text.IO as T
 
 import System.IO
 
 import Control.Applicative
-import Text.Printf (printf)
 import Text.Read (readEither)
 
 import Control.Monad (unless, when)
@@ -38,13 +33,99 @@ import Control.Monad.Reader.Class (MonadReader, asks)
 import Lens.Micro
 import Lens.Micro.Extras
 
-import qualified Toml hiding (parse)
 import qualified Toml.Parser.Core as Toml (eof, errorBundlePretty, parse)
 import qualified Toml.Parser.Value as Toml
 
 import Data.Time
 
 import System.Directory
+
+
+-----------------------------------------
+-- getters and setter for ConfigValue
+
+getBool :: ConfigValue 'TBool -> Bool
+getBool (BoolV bool) = bool
+
+setBool :: ConfigValue 'TBool -> Bool -> ConfigValue 'TBool
+setBool _ = BoolV
+
+getDefaultBookmark :: ConfigValue 'TDefaultBookmark -> DefaultBookmark
+getDefaultBookmark (DefaultBookmarkV bm) = bm
+
+setDefaultBookmark :: ConfigValue 'TDefaultBookmark -> DefaultBookmark -> ConfigValue 'TDefaultBookmark
+setDefaultBookmark _ = DefaultBookmarkV
+
+getCommand :: ConfigValue 'TCommand -> T.Text
+getCommand (CommandV t) = t
+
+setCommand :: ConfigValue 'TCommand -> T.Text -> ConfigValue 'TCommand
+setCommand _ = CommandV
+
+getList :: ConfigValue ('TList t) -> [ConfigValue t]
+getList (ListOfV xs) = xs
+
+setList :: ConfigValue ('TList t) -> [ConfigValue t] -> ConfigValue ('TList t)
+setList _ = ListOfV
+
+getMaybe :: ConfigValue ('TMaybe t) -> Maybe (ConfigValue t)
+getMaybe (MaybeV val) = val
+
+setMaybe :: ConfigValue ('TMaybe t) -> Maybe (ConfigValue t) -> ConfigValue ('TMaybe t)
+setMaybe _ = MaybeV
+
+-----------------------------------------
+-- Lenses for Config
+
+failOnError :: Lens' Config Bool
+failOnError = lens getter setter
+  where
+    getter = getBool . _failOnError
+    setter cfg bool = cfg { _failOnError = setBool (_failOnError cfg) bool }
+
+displayCreationTime :: Lens' Config Bool
+displayCreationTime = lens getter setter
+  where
+    getter = getBool . _displayCreationTime
+    setter cfg bool = cfg { _displayCreationTime = setBool (_displayCreationTime cfg) bool }
+
+enableClearing :: Lens' Config Bool
+enableClearing = lens getter setter
+  where
+    getter = getBool . _enableClearing
+    setter cfg bool = cfg { _enableClearing = setBool (_enableClearing cfg) bool }
+
+enableReset :: Lens' Config Bool
+enableReset = lens getter setter
+  where
+    getter = getBool . _enableReset
+    setter cfg bool = cfg { _enableReset = setBool (_enableReset cfg) bool }
+
+backupBeforeClear :: Lens' Config Bool
+backupBeforeClear = lens getter setter
+  where
+    getter = getBool . _backupBeforeClear
+    setter cfg bool = cfg { _backupBeforeClear = setBool (_backupBeforeClear cfg) bool }
+
+defaultBookmarks :: Lens' Config [DefaultBookmark]
+defaultBookmarks = lens getter setter
+  where
+    getter :: Config -> [DefaultBookmark]
+    getter = map getDefaultBookmark . getList . _defaultBookmarks
+
+    setter :: Config -> [DefaultBookmark] -> Config
+    setter cfg bms = cfg { _defaultBookmarks = setList (_defaultBookmarks cfg) (fmap DefaultBookmarkV bms)}
+
+defaultCommand :: Lens' Config (Maybe T.Text)
+defaultCommand = lens getter setter
+  where
+    getter :: Config -> Maybe T.Text
+    getter = fmap getCommand . getMaybe . _defaultCommand
+
+    setter :: Config -> Maybe T.Text -> Config
+    setter cfg cmd = cfg { _defaultCommand = setMaybe (_defaultCommand cfg) (fmap CommandV cmd) }
+
+-----------------------------------------
 
 -- | A version of the lens "use" function for 'MonadReader'.
 asks' :: MonadReader a m => SimpleGetter a b -> m b
@@ -68,80 +149,6 @@ assertVerbose err check = do
   res <- check
   when (shouldFail && not res) $ throwError err
   return res
-
--- | Get a list of the pairs in a Toml hashmap.
-pairsToKeyVals :: HashMap.HashMap Toml.Key Toml.AnyValue -> [(Toml.Key, Toml.AnyValue)]
-pairsToKeyVals = HashMap.toList
-
--- | Get a list of the pairs in a prefixmap of Toml tables.
-tablesToKeyVals :: Toml.PrefixMap Toml.TOML -> [(Toml.Key, Toml.AnyValue)]
-tablesToKeyVals = concatMap helper . Toml.toList
-  where
-    helper (k, toml) = [(k <> k2, v) | (k2, v) <- tomlToKeyVals toml]
-
--- | Get a list of the pairs in a prefixmap of Toml tables arrays.
-tableArraysToKeyVals :: HashMap.HashMap Toml.Key (NE.NonEmpty Toml.TOML) -> [(Toml.Key, Toml.AnyValue)]
-tableArraysToKeyVals = concatMap helper . HashMap.toList
-  where
-    helper :: (Toml.Key, NE.NonEmpty Toml.TOML) -> [(Toml.Key, Toml.AnyValue)]
-    helper (k, ne) = [(k <> k2, v) | toml <- NE.toList ne, (k2, v) <- tomlToKeyVals toml]
-
--- | Convert a 'Toml.TOML' object to a list of keys and values.
-tomlToKeyVals :: Toml.TOML -> [(Toml.Key, Toml.AnyValue)]
-tomlToKeyVals toml =
-  pairsToKeyVals (Toml.tomlPairs toml)
-  <> tablesToKeyVals (Toml.tomlTables toml)
-  -- <> tableArraysToKeyVals (Toml.tomlTableArrays toml) -- TODO
-
--- | Text representation of a Toml value. Copied directly from 'Toml.Type.Printer' source code.
-valText :: Toml.Value t -> T.Text
-valText (Toml.Bool b)    = T.toLower $ tshow b
-valText (Toml.Integer n) = tshow n
-valText (Toml.Double dub)  = showDouble dub
-  where
-    showDouble :: Double -> T.Text
-    showDouble d | isInfinite d && d < 0 = "-inf"
-                 | isInfinite d = "inf"
-                 | isNaN d = "nan"
-                 | otherwise = tshow d
-valText (Toml.Text s)    = showTextUnicode s
-  where
-    showTextUnicode :: T.Text -> T.Text
-    showTextUnicode text = T.pack $ show finalText
-      where
-        xss = T.unpack text
-        finalText = foldl' (\acc (ch, asciiCh) -> acc ++ getCh ch asciiCh) "" asciiArr
-
-        asciiArr = zip xss $ asciiStatus xss
-
-        getCh :: Char -> Bool -> String
-        getCh ch True  = [ch] -- it is true ascii character
-        getCh ch False = printf "\\U%08x" (ord ch) :: String -- it is not true ascii character, it must be encoded
-
-        asciiStatus :: String -> [Bool]
-        asciiStatus = map isAscii
-valText (Toml.Zoned zTime)   = showZonedTime zTime
-  where
-    showZonedTime :: ZonedTime -> T.Text
-    showZonedTime t = T.pack $ showZonedDateTime t <> showZonedZone t
-      where
-        showZonedDateTime = formatTime defaultTimeLocale "%FT%T%Q"
-        showZonedZone
-            = (\(x,y) -> x ++ ":" ++ y)
-            . (\z -> splitAt (length z - 2) z)
-            . formatTime defaultTimeLocale "%z"
-valText (Toml.Local l)   = tshow l
-valText (Toml.Day d)     = tshow d
-valText (Toml.Hours h)   = tshow h
-valText (Toml.Array arr)   = withLines Toml.defaultOptions valText arr
-  where
-    withLines :: Toml.PrintOptions -> (Toml.Value t -> T.Text) -> [Toml.Value t] -> T.Text
-    withLines Toml.PrintOptions{..} valTxt a = case printOptionsLines of
-        Toml.OneLine -> "[" <> T.intercalate ", " (map valTxt a) <> "]"
-        Toml.MultiLine -> off <> "[ " <> T.intercalate (off <> ", ") (map valTxt a) <> off <> "]"
-      where
-        off :: T.Text
-        off = "\n" <> stimes printOptionsIndent " "
 
 -- | Given a file name and an extension, try to find a suitable path for
 -- backing up that file. Used by 'backupFile'.
@@ -208,7 +215,7 @@ formatBookmark shouldDisplayTime indexWidth (Bookmark dir idx zTime mbName) =
   let num = T.justifyRight indexWidth ' ' $ tshow idx
       timeStr = T.pack $ formatTime defaultTimeLocale "%D %T" zTime
       d = case mbName of Nothing    -> dir
-                         Just name  -> dir <> " " <> name
+                         Just name  -> dir <> "\t(" <> name <> ")"
 
   in if shouldDisplayTime
      then num <> ". " <> timeStr <> "\t" <> d
@@ -225,6 +232,25 @@ formatBookmarks shouldDisplayTime bms = map (formatBookmark shouldDisplayTime in
   where
     indexWidth = maximumDefault 1 $ map (length . show . view bookmarkIndex) bms
 
+-- | Format a 'DefaultBookmark'. Used for the "config print" command and error reporting
+-- during other commands.
+--
+-- @formatDefaultBookmark bm@ returns a pretty representation of @bm@.
+formatDefaultBookmark :: DefaultBookmark -> T.Text
+formatDefaultBookmark (DefaultBookmark dir mbName) =
+  case mbName of Nothing    -> dir
+                 Just name  -> dir <> "\t(" <> name <> ")"
+
 -- | Show a value as a 'T.Text' instead of a 'String'.
 tshow :: Show a => a -> T.Text
 tshow = T.pack . show
+
+-- | Format a config value. Used for the "config print" command.
+formatConfigValue :: AnyConfigValue -> T.Text
+formatConfigValue (AnyConfigValue (BoolV bool)) = tshow bool
+formatConfigValue (AnyConfigValue (DefaultBookmarkV bm)) = formatDefaultBookmark bm
+formatConfigValue (AnyConfigValue (CommandV t)) = tshow t
+formatConfigValue (AnyConfigValue (MaybeV t)) = case t of
+                                                  Nothing -> ""
+                                                  Just t' -> formatConfigValue (AnyConfigValue t')
+formatConfigValue (AnyConfigValue (ListOfV xs)) = T.intercalate "\n" (["["] <> map (("  " <>) . formatConfigValue . AnyConfigValue) xs <> ["]"])