+++ CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to the
 - Integrate [headroom](https://github.com/vaclavsvejcar/headroom).
 - Some util functions for testing in HoYo/Test/HoYo.hs.
 - Add more bookmark + env tests.
+- Better pretty printing in `hoyo config print` command.
+- Stricter config type safety using GADTs parameterised by `ConfigValueType`.
+- Arbitrary values for `ConfigValue`.
 
 ### Fixed
 
@@ -23,8 +26,11 @@ and this project adheres to the
 ### Changed
 
 - Refactored uniqueness tests in `hoyo add` to their own functions.
+- `Config` now uses `ConfigValue`s instead of raw `Bool`/`Text` etc.
 
 ### Removed
+
+- All the auxiliary pretty-printing functions that operated on Toml values.
 
 ## [0.2.3.1] - 2023-01-21